fix: context error when log out
- only refetch notes and labels if `currentUser` is not null
- add light mode to `SettingCorner` buttons

+++ src/contexts/AllLabelsContext.tsx
@@ -12,43 +12,45 @@ export default function AllLabelsContextProvider({ children }: { children: JSX.E
   const currentUser = useContext(UserContext) as User;
   const [allLabels, setAllLabels] = useState<LabelDbData[]>([]);
   useEffect(() => {
-    const fetchLabels = async () => {
-      const fetchResult = await getAllLabels(currentUser.uid);
-      if (fetchResult) setAllLabels(fetchResult);
-    };
-    fetchLabels();
-    supabase
-      .channel('labels-changes')
-      .on(
-        'postgres_changes',
-        {
-          event: '*',
-          schema: 'public',
-          table: 'labels',
-          filter: `user_id=eq.${currentUser.uid}`,
-        },
-        (_payload) => fetchLabels(), // listen for changes, refetch if there is one
-      )
-      .on(
-        'postgres_changes',
-        {
-          event: 'DELETE',
-          schema: 'public',
-          table: 'notes_labels',
-          filter: `user_id=eq.${currentUser.uid}`,
-        },
-        async (payload) => {
-          const deletedLabelId = payload.old.label_id;
-          const notesLabelsWithThisLabelId = await getNotesLabels(deletedLabelId);
-          if (notesLabelsWithThisLabelId && notesLabelsWithThisLabelId.length === 0) {
-            // if there's no row in `notes_labels` with this label_id
-            // then there is no note with this label
-            // delete label from `labels`
-            await deleteLabelById(deletedLabelId);
-          }
-        },
-      )
-      .subscribe();
+    if (currentUser) {
+      const fetchLabels = async () => {
+        const fetchResult = await getAllLabels(currentUser.uid);
+        if (fetchResult) setAllLabels(fetchResult);
+      };
+      fetchLabels();
+      supabase
+        .channel('labels-changes')
+        .on(
+          'postgres_changes',
+          {
+            event: '*',
+            schema: 'public',
+            table: 'labels',
+            filter: `user_id=eq.${currentUser.uid}`,
+          },
+          (_payload) => fetchLabels(), // listen for changes, refetch if there is one
+        )
+        .on(
+          'postgres_changes',
+          {
+            event: 'DELETE',
+            schema: 'public',
+            table: 'notes_labels',
+            filter: `user_id=eq.${currentUser.uid}`,
+          },
+          async (payload) => {
+            const deletedLabelId = payload.old.label_id;
+            const notesLabelsWithThisLabelId = await getNotesLabels(deletedLabelId);
+            if (notesLabelsWithThisLabelId && notesLabelsWithThisLabelId.length === 0) {
+              // if there's no row in `notes_labels` with this label_id
+              // then there is no note with this label
+              // delete label from `labels`
+              await deleteLabelById(deletedLabelId);
+            }
+          },
+        )
+        .subscribe();
+    }
   }, [currentUser]);
 
   return <AllLabelsContext.Provider value={allLabels}>{children}</AllLabelsContext.Provider>;
